requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.2.1
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
Django==3.1.7
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.7.2
incremental==21.3.0
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
msgpack==1.0.2
Pillow==8.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==20.0.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
zipp==3.4.1
zope.interface==5.2.0
